requirements.txt
mss
pytesseract
requests
Pillow
